Move `plutono`-related functions in dedicated file
