+++ pkg/operation/botanist/plutono.go
@@ -1,0 +1,95 @@
+// Copyright 2023 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package botanist
+
+import (
+	"context"
+	"fmt"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/pointer"
+
+	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
+	"github.com/gardener/gardener/pkg/component"
+	"github.com/gardener/gardener/pkg/component/plutono"
+	"github.com/gardener/gardener/pkg/component/shared"
+	"github.com/gardener/gardener/pkg/features"
+	"github.com/gardener/gardener/pkg/operation/common"
+	gardenerutils "github.com/gardener/gardener/pkg/utils/gardener"
+	kubernetesutils "github.com/gardener/gardener/pkg/utils/kubernetes"
+)
+
+// DefaultPlutono returns a deployer for Plutono.
+func (b *Botanist) DefaultPlutono() (plutono.Interface, error) {
+	var wildcardCertName *string
+	if b.ControlPlaneWildcardCert != nil {
+		wildcardCertName = pointer.String(b.ControlPlaneWildcardCert.GetName())
+	}
+
+	return shared.NewPlutono(
+		b.SeedClientSet.Client(),
+		b.Shoot.SeedNamespace,
+		b.ImageVector,
+		b.SecretsManager,
+		"",
+		component.ClusterTypeShoot,
+		b.ComputePlutonoHost(),
+		b.ShootUsesDNS(),
+		features.DefaultFeatureGate.Enabled(features.MachineControllerManagerDeployment),
+		b.Shoot.NodeLocalDNSEnabled,
+		b.Shoot.IsWorkerless,
+		b.Shoot.VPNHighAvailabilityEnabled,
+		v1beta1constants.PriorityClassNameShootControlPlane100,
+		b.Shoot.GetReplicas(1),
+		wildcardCertName,
+		b.Shoot.WantsVerticalPodAutoscaler,
+	)
+}
+
+// DeployPlutono deploys the plutono in the Seed cluster.
+func (b *Botanist) DeployPlutono(ctx context.Context) error {
+	// disable monitoring if shoot has purpose testing or monitoring and vali is disabled
+	if !b.Operation.WantsPlutono() {
+		if err := b.Shoot.Components.ControlPlane.Plutono.Destroy(ctx); err != nil {
+			return err
+		}
+
+		secretName := gardenerutils.ComputeShootProjectSecretName(b.Shoot.GetInfo().Name, gardenerutils.ShootProjectSecretSuffixMonitoring)
+		return kubernetesutils.DeleteObject(ctx, b.GardenClient, &corev1.Secret{ObjectMeta: metav1.ObjectMeta{Name: secretName, Namespace: b.Shoot.GetInfo().Namespace}})
+	}
+
+	// TODO(rickardsjp, istvanballok): Remove in release v1.77 once the Grafana to Plutono migration is complete.
+	if err := common.DeleteGrafana(ctx, b.SeedClientSet, b.Shoot.SeedNamespace); err != nil {
+		return err
+	}
+
+	if err := b.Shoot.Components.ControlPlane.Plutono.Deploy(ctx); err != nil {
+		return err
+	}
+
+	credentialsSecret, found := b.SecretsManager.Get(v1beta1constants.SecretNameObservabilityIngressUsers)
+	if !found {
+		return fmt.Errorf("secret %q not found", v1beta1constants.SecretNameObservabilityIngressUsers)
+	}
+
+	return b.syncShootCredentialToGarden(
+		ctx,
+		gardenerutils.ShootProjectSecretSuffixMonitoring,
+		map[string]string{v1beta1constants.GardenRole: v1beta1constants.GardenRoleMonitoring},
+		map[string]string{"url": "https://" + b.ComputePlutonoHost()},
+		credentialsSecret.Data,
+	)
+}
